Add option for quick connect logins
This change changes the login flow to first request the uri of the server, then the user selects login with password or quick connect. At that point, the existing password login flow takes over, or a new one that gets a quick connect code is used.

As part of this change, the auth information stored by the system has been changed. Previously the password had been stored. This is probably not needed (at least basic usage shows it doesn't affect anything) since there's a refresh token there. Since both login flows provide a refresh token, we can store that instead.